pyproject.toml
[tool.poetry]
name = "pip"
version = "0.1.0"
description = ""
authors = ["Suyash Mohanty"]
readme = "README.md"



[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.0"
uvicorn = "^0.30.5"
sqlalchemy = "^2.0.32"
pyodbc = "^5.1.0"
pydantic = "^2.8.2"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
openai = "1.40.0"
python-dotenv = "^1.0.1"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.9"
ipykernel = "^6.29.5"
langchain-community = "^0.2.14"
bs4 = "^0.0.2"
langchain = "^0.2.15"
beautifulsoup4 = "^4.12.3"
langchain-openai = "^0.1.23"
faiss-cpu = "^1.8.0.post1"
pypdf = "^4.3.1"
pymssql = "^2.3.1"
streamlit = "^1.38.0"
langchain-groq = "^0.1.9"
mysql-connector-python = "^9.0.0"
langchain-chroma = "^0.1.3"
langchain-core = "^0.2.38"
pymupdf = "^1.24.10"
langchain-text-splitters = "^0.2.4"
sentence-transformers = "^3.0.1"
requests = "^2.32.3"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
